Make targets/target shorthands available on every actions context
